LOCAL

+++ src/app/shopping-cart/shopping-cart.ts
@@ -55,16 +55,29 @@ export class ShoppingCart implements OnInit {
     this.loading.set(false);
   }
   removeFromCart(purchaseId: number, productId: number) {
+  if (this.customerId) {
+    // משתמש מחובר → שלח לשרת
     this.purchaseDetailsService.RemoveProduct(purchaseId, productId).subscribe({
       next: (res) => {
-        console.log('המוצר נמחק בהצלחה', res);
+        console.log('המוצר נמחק בהצלחה מהשרת', res);
         this.loadProductsCart();
       },
       error: (err) => {
-        console.error('שגיאה במחיקת המוצר:', err);
+        console.error('שגיאה במחיקת המוצר מהשרת:', err);
       }
     });
+  } else {
+    // משתמש לא מחובר → הסרה מה-localStorage
+    const localCartRaw = localStorage.getItem('cart');
+    if (localCartRaw) {
+      let localCart: PurchaseDetail[] = JSON.parse(localCartRaw);
+      localCart = localCart.filter(item => item.productId !== productId);
+      localStorage.setItem('cart', JSON.stringify(localCart));
+      this.list.set(localCart); // עדכון ה-UI
+      console.log('המוצר נמחק מהעגלה המקומית', productId);
+    }
   }
+}
 
 
 }
